% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab6Function.R
\name{mybin}
\alias{mybin}
\title{Lab 5 Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{x}{A vector of quantitative data}
}
\value{
A vector of squared components
}
\description{
Lab 5 Function
}
\examples{


}
